+++ jsonype/__init__.py
@@ -27,6 +27,8 @@ from .basic_to_json_converters import FromNone as FromNone  # noqa: W0611
 from .basic_to_json_converters import FromSequence as FromSequence  # noqa: W0611
 from .basic_to_json_converters import FromSimple as FromSimple  # noqa: W0611
 from .basic_to_json_converters import ToJsonConverter as ToJsonConverter  # noqa: W0611
+from .dataclass_converters import FromDataclass as FromDataclass  # noqa: W0611
+from .dataclass_converters import ToDataclass as ToDataclass  # noqa: W0611
 from .named_tuple_converters import FromNamedTuple as FromNamedTuple  # noqa: W0611
 from .named_tuple_converters import ToNamedTuple as ToNamedTuple  # noqa: W0611
 from .typed_json import TypedJson as TypedJson  # noqa: W0611