+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+## next
+
+- Add support for converting from/to dataclasses
+
 ## 0.2.0
 
 - Add support for converting from/to NamedTuples